+++ src/acconeer/exptool/app/new/ui/main_window.py
@@ -31,14 +31,7 @@ from acconeer.exptool.app.new.app_model import AppModel
 
 from .connection_widget import ClientConnectionWidget, GenerationSelection
 from .flash_widget import FlashButton
-from .misc import (
-    BUTTON_ICON_COLOR,
-    ConnectionHint,
-    ExceptionWidget,
-    InaccessibleDeviceHint,
-    UnflashedDeviceHint,
-    VerticalSeparator,
-)
+from .misc import BUTTON_ICON_COLOR, ExceptionWidget, HintWidget, VerticalSeparator
 from .plugin_widget import PluginControlArea, PluginPlotArea, PluginSelection
 from .recording_widget import RecordingWidget
 from .utils import ScrollAreaDecorator, TopAlignDecorator
@@ -92,9 +85,7 @@ class TopBar(QFrame):
         self.layout().addWidget(ClientConnectionWidget(app_model, self))
         self.layout().addWidget(VerticalSeparator(self))
         self.layout().addWidget(FlashButton(app_model, self))
-        self.layout().addWidget(ConnectionHint(app_model, self))
-        self.layout().addWidget(UnflashedDeviceHint(app_model, self))
-        self.layout().addWidget(InaccessibleDeviceHint(app_model, self))
+        self.layout().addWidget(HintWidget(app_model, self))
         self.layout().addStretch(1)
         self.layout().addWidget(RecordingWidget(app_model, self))